Move IBlobPersistor out of Modules/
