+++ vnext/Shared/IBlobPersistor.h
